+++ fraisier/__init__.py
@@ -17,5 +17,5 @@ Usage:
     fraisier status <fraise> <environment>  # Check fraise status
 """
 
-__version__ = "0.2.4"
+__version__ = "0.3.0"
 __all__ = ["__version__"]